Develop

+++ client/src/components/TableAdmin.jsx
@@ -1,12 +1,11 @@
 import React, { useState } from 'react';
 import { useDispatch, useSelector } from "react-redux";
 import { useEffect } from 'react';
-import {useNavigate} from "react-router-dom";
 import {getFoods, getMenus, addFoodToMenu, deleteFood} from "../redux/actions";
+import swal from 'sweetalert';
 
 const TableAdmin = () => {
     const dispatch = useDispatch();
-    const navigate = useNavigate();
     const foods = useSelector((state) => state.foods);
     const menuus = useSelector((state) => state.menus);
     useEffect(() => {
@@ -35,33 +34,61 @@ const TableAdmin = () => {
     function handleAdd(e, foood) {
         if(menu !== "") {
             if(findFoodMenu(foood)) {
-                alert("This food is already in this menu");
+                swal({
+                    title: "La comida ya esta en el menu",
+                    icon: "warning",
+                    button: "Aceptar",
+                });
+                dispatch(getFoods());
+                dispatch(getMenus());
             } else {
             dispatch(addFoodToMenu({
                 menu : menu,
                 food : foood
             }));
-            alert("Food added to menu");
-            navigate("/");
+            swal({
+                title: "Comida agregada al menu",
+                icon: "success",
+                button: "Aceptar",
+            });
+            dispatch(getFoods());
+            dispatch(getMenus());
         } 
      } else {
-            alert("Please select a menu");
+            swal({
+                title: "Debe seleccionar un menu",
+                icon: "warning",
+                button: "Aceptar",
+            });
         }
     }
 
     function handleDelete(e, id) {
+        swal({
+            title: "Esta seguro que desea eliminar la comida?",
+            icon: "warning",
+            buttons: ["Cancelar", "Aceptar"],
+            dangerMode: true,
+        }).then((willDelete) => {
+        if(willDelete) {
         dispatch(deleteFood(id));
-        alert("Food deleted");
-        navigate("/");
+        swal({
+            title: "Comida eliminada",
+            icon: "success",
+            button: "Aceptar",
+        });
+        // alert("Comida eliminada");
+        dispatch(getFoods());
+        dispatch(getMenus());
+        } else {
+            swal({
+                title: "Comida no eliminada",
+                icon: "success",
+                button: "Aceptar",
+            });
+        }
+        });
     }
-
-
-
-
-
-
-
-
 
     
     return (
@@ -78,10 +105,10 @@ const TableAdmin = () => {
                             <th className="text-left p-3 px-5">Food</th>
                             <th className="text-left p-3 px-5">View Menu</th>
                             <th className="text-left p-3 px-5">
-                            <select className="bg-transparent" onChange={(e)=> handleChange(e)}>
-                                            <option hidden select>Select Menu</option>
+                            <select className="bg-transparent" onChange={(e)=> handleChange(e)} defaultValue="Menu"  >
+                                            <option hidden>Select Menu</option>
                                                 {
-                                                    menuus.map((menu) => {
+                                                    menuus?.map((menu) => {
                                                         return(
                                                             <option key={menu.id} value={menu.name}>{menu.name}</option>
                                                         )
@@ -98,27 +125,24 @@ const TableAdmin = () => {
                                         <td className="p-3 px-5">{food.name}</td>
                                         <td className="p-3 px-5">
                                             <select className="bg-transparent">
-                                    
+                                            <option hidden>View Menus</option>
                                                 
-                                            {
+                                            {/* {
                                                 food.menus?.map((menu) => {
                                                     return(
-                                                        <option key={menu.id} value={menu.name} disabled selected>{menu.name}</option>
+                                                        <option key={menu.id} value={menu.name} disabled>{menu.name}</option>
                                                     )
                                                 })
-                                            }
-                                            <option selected disabled hidden>View Menus</option>
+                                            } */}
+                                            
                                             </select>
                                         </td>
                                         <td className="p-3 px-5">{menu}
                                         </td>
                                         <td className="p-3 px-5 flex justify-end"><button type="button" onClick={(e)=> handleAdd(e, food.name)} className="mr-3 text-sm bg-blue-500 hover:bg-blue-700 text-white py-1 px-2 rounded focus:outline-none focus:shadow-outline">Save Menu</button><button type="button" onClick={(e) => handleDelete(e, food.id)} className="text-sm bg-red-500 hover:bg-red-700 text-white py-1 px-2 rounded focus:outline-none focus:shadow-outline">Delete Food</button></td>
                                     </tr>
-
                                 )
-
                             })
-                
                         }
                     </tbody>
                 </table>
